feat: provision an opensearch instance
Rather than rolling our own ELK stack, we decided to opt for the AWS-hosted service, which now uses
OpenSearch, which is the new open source version of Elastic. Sending the telemetry information from
the node to OpenSearch requires two services: the OpenTelemetry data collector and the Data Prepper.
We host these on ECS rather than using EC2 instances.

There are actually three different Terraform runs here. There is an explanation for this in the
README, which you can see in the diff.

The whole process is not completely automated, but it would definitely be possible to do it at some
later stage.