feat: deploy open telemetry collector
The Telemetry Collector is also deployed as an ECS service.

The Data Prepper is registered with service discovery so that the Telemetry Collector can refer to
it by a DNS name. The ECS infrastructure takes care of associating the replicated ECS tasks with
private DNS records in Route 53.

Each service is deployed in its own security group and all the network ACLs are defined. This
involves defining all the inbound and outbound rules explicitly.