feat: run services using service connect
This was a setup that used the ECS 'Service Connect' service for simple service discovery. It seemed
like the Telemetry Collector application was not compatible with this setup, but I'd like to have it
preserved because it may be useful at some other point.

I'm going to keep the setup on this branch.