chore: new data prepper and telemetry collector config
The documentation for the Data Prepper was very recently updated to provide new default values for
the buffer settings. These seem to make the Telemetry Collector and the Data Prepper work better
together.

The Telemetry Collector was also updated to use a batch processor, which is recommended in the docs
for trace analytics:
https://github.com/open-telemetry/opentelemetry-collector/tree/main/processor#recommended-processors