refactor: move code to opensearch specific modules
I now want to include more than just Opensearch in this repository, so I'm moving the Opensearch
related things to their own modules.

There is a base networking module which will include the VPC where the testnet instances will be
launched. The Opensearch stuff will be removed from that.

The Makefile was also converted to a Justfile for reduced duplication.